requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
motor==3.7.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pymongo==4.9.1
starlette~=0.46.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.1
python-multipart>=0.0.5